requirements.txt
fastapi
ffmpeg-python
opencv-contrib-python
opencv-python
Pillow
pydantic_settings
requests
uvicorn